requirements.txt
asgiref==3.5.0
psycopg2-binary==2.9.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
jsonschema==4.4.0
Markdown==3.3.6
MarkupSafe==2.1.1
openapi-codec==1.3.2
PyJWT==2.3.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.7.0
